+++ js/views/FailView.js
@@ -1,30 +1,28 @@
 import content from '../data/game-content';
-import {calculatePoints} from '../data/calc-points';
-import {showResults} from '../data/show-results';
+import {showResults} from '../game/show-results';
 import AbstractView from '../views/AbstractView';
+import {InitialState} from '../data/game-data';
 
 export default class FailView extends AbstractView {
   constructor(state) {
     super();
     this.state = state;
-    const {points} = calculatePoints(this.state);
-    this.points = points;
+
     this.result = {
-      UserPoint: points,
-      time: state.time,
-      lives: state.lives
+      userPoint: null,
+      time: InitialState.time - this.state.time,
+      lives: this.state.lives
     };
-    this.results = [10, 5, 6, 7];
-    this.replic = state.lives < 0 ? `${content.result.loseReplic}` : `${content.result.timeLoseReplic}`;
+    this.replica = this.state.lives <= 0 ? `${content.result.loseReplica}` : `${content.result.timeLoseReplica}`;
   }
 
   get template() {
     return `
       <section class="result">
-        <div class="result__logo"><img src="img/melody-logo.png" alt="${content.header.logo}" width="186" height="83"></div>
-        <h2 class="result__title">${this.replic}</h2>
-        <p class="result__total result__total--fail">${showResults(this.results, this.result)}</p>
-        <button class="result__replay" type="button">${content.buttons.loseReplay}</button>
+      <div class="result__logo"><img src="img/melody-logo.png" alt="${content.header.logo}" width="186" height="83"></div>
+      <h2 class="result__title">${this.replica}</h2>
+      <p class="result__total result__total--fail">${showResults([], this.result)}</p>
+      <button class="result__replay" type="button">${content.buttons.loseReplay}</button>
       </section>
     `;
   }
